Final Touchs Before Front-End

+++ Developers/adham.py
@@ -1,124 +1,95 @@
-from flask import Blueprint, Flask, jsonify, request
-import mysql.connector
+from flask import Blueprint, jsonify, request
+from DB_Connect import myDB
 
 
-app = Flask(__name__)
-
-# MySQL configuration
-myDB = mysql.connector.connect(
-    host='mysql-158141-0.cloudclusters.net',
-    port=10014,
-    user='admin',
-    password='oGFpwVr8',
-    database='GroceryShop'
-)
+app = Blueprint('adham', __name__)
 
 # Endpoint to search products by keyword
 @app.route('/getBySearch', methods=['GET'])
 def get_by_search():
     try:
         keyword = request.args.get('keyword')
-
+        # Check Parameter Existence
+        if keyword == None:
+            return {"success": False, "message":"Missing Parameter"},406
+        
         cursor = myDB.cursor(dictionary=True)
-
-
-
-       
-
-        query = f"SELECT * FROM product WHERE productName LIKE '%{keyword}%'"
+        query = f"""
+                SELECT
+                    product.id,
+                    product.productName,
+                    product.productPrice,
+                    product.productImage,
+                    product.quantity,
+                    product.calories,
+                    product.discount,
+                    Brand.name as brand,
+                    Brand.nationality 
+                FROM product INNER JOIN Brand ON product.brand = Brand.id WHERE product.productName LIKE '%{keyword}%';"""
         cursor.execute(query)
         products = cursor.fetchall()
 
-        product_list = [
-            {
-                'id': product['id'],
-                'productImage': product['productImage'],
-                'productName': product['productName'],
-                'productPrice': float(product['productPrice']),
-                'quantity': product['quantity'],
-                'nationality': product['nationality'],
-                'brand': product['brand'],
-                'discount': float(product['discount']),
-                'calories': product['calories']
-            }
-            for product in products
-        ]
-
         cursor.close()
-
-        return jsonify({'success': True, 'products': product_list})
+        return jsonify({'success': True, 'data': products})
 
     except Exception as e:
         print('Error executing search query:', str(e))
-        return jsonify({'success': False, 'error': 'Internal Server Error'}), 500
+        return jsonify({'success': False, 'message': f'Internal Server Error: {str(e)}'}), 500
 
 # Endpoint to filter products by nationality
 @app.route('/filterByNationality', methods=['GET'])
 def filter_by_nationality():
     try:
-        nationality = request.args.get('Nationality', '')
-
+        nationality = request.args.get('nationality')
+        if nationality == None:
+            return {"success": False, "message":"Missing Parameter"},406
         cursor = myDB.cursor(dictionary=True)
 
-        query = f"SELECT * FROM product WHERE nationality = '{nationality}'"
+        query = f"""
+                SELECT
+                    product.id,
+                    product.productName,
+                    product.productPrice,
+                    product.productImage,
+                    product.quantity,
+                    product.calories,
+                    product.discount,
+                    Brand.name as brand,
+                    Brand.nationality 
+                FROM product INNER JOIN Brand ON product.brand = Brand.id WHERE Brand.nationality = "{nationality}";"""
         cursor.execute(query)
         products = cursor.fetchall()
 
-        product_list = [
-            {
-                'id': product['id'],
-                'productImage': product['productImage'],
-                'productName': product['productName'],
-                'productPrice': float(product['productPrice']),
-                'quantity': product['quantity'],
-                'nationality': product['nationality'],
-                'brand': product['brand'],
-                'discount': float(product['discount']),
-                'calories': product['calories']
-            }
-            for product in products
-        ]
-
         cursor.close()
 
-        return jsonify({'success': True, 'products': product_list})
+        return jsonify({'success': True, 'data': products})
 
     except Exception as e:
         print('Error executing nationality filter query:', str(e))
-        return jsonify({'success': False, 'error': 'Internal Server Error'}), 500
+        return jsonify({'success': False, 'message': f'Internal Server Error: {str(e)}'}), 500
 
 # Endpoint to get payment methods for a user
 @app.route('/getPaymentMethods', methods=['GET'])
 def get_payment_methods():
     try:
-        user_id = request.args.get('userId', '')
-
+        user_id = request.args.get('userId')
+        if user_id == None:
+            return {"success": False, "message":"Missing Parameter"},406
         cursor = myDB.cursor(dictionary=True)
 
         query = f"SELECT * FROM PayementMethods WHERE userID = '{user_id}'"
         cursor.execute(query)
         payment_methods = cursor.fetchall()
 
-        payment_methods_list = [
-            {
-                'id': method['id'],
-                'userID': method['userID'],-
-                'cardNumber': method['cardNumber'],
-                'cvv': method['cvv'],
-                'cardHolderName': method['cardHolderName'],
-            }
-            for method in payment_methods
-        ]
-
         cursor.close()
 
-        return jsonify({'success': True, 'paymentMethods': payment_methods_list})
+        return jsonify({'success': True, 'data': payment_methods})
 
     except Exception as e:
         print('Error executing getPaymentMethods query:', str(e))
-        return jsonify({'success': False, 'error': 'Internal Server Error'}), 500
-    
- # Endpoint to add payment methods
+        return jsonify({'success': False, 'message': f'Internal Server Error: {str(e)}'}), 500
+
+# Endpoint to add payment methods
 @app.route('/addPaymentMethods', methods=['POST'])
 def add_payment_methods():
     try:
@@ -129,6 +100,13 @@ def add_payment_methods():
         cvv = data.get('cvv')
         card_holder_name = data.get('cardHolderName')
 
+        # Check Parameter Existence
+        if (user_id == None
+        or card_number == None
+        or cvv == None
+        or card_holder_name == None):
+            return {"success": False, "message":"Missing Parameter"},406
+ 
         cursor = myDB.cursor()
         # Corrected table name to 'PaymentMethods'
         query = "INSERT INTO PayementMethods (userID, cardNumber, cvv, cardHolderName) VALUES (%s, %s, %s, %s)"
@@ -140,15 +118,17 @@ def add_payment_methods():
         return jsonify({'success': True, 'message': 'Payment method added successfully'})
     except Exception as e:
         print('Error adding payment method:', str(e))
-        return jsonify({'success': False, 'error': 'Internal Server Error'}), 500
+        return jsonify({'success': False, 'message': f'Internal Server Error: {str(e)}'}), 500
 
 
 # Endpoint to remove payment methods
 @app.route('/removePaymentMethods', methods=['DELETE'])
 def remove_payment_methods():
     try:
-        user_id = request.args.get('userId', '')
-        payment_method_id = request.args.get('paymentMethodId', '')
+        user_id = request.args.get('userId')
+        payment_method_id = request.args.get('paymentMethodId')
+        if user_id == None or payment_method_id == None:
+            return jsonify({"success": False, "message":"Missing Parameter"}),406
         cursor = myDB.cursor()
         # Delete the specified payment method
         query = "DELETE FROM PayementMethods WHERE userID = %s AND id = %s"
@@ -161,14 +141,17 @@ def remove_payment_methods():
         return jsonify({'success': True, 'message': 'Payment method removed'}), 500
     except Exception as e:
         # Handle the exception (you might want to log it or return an error response)
-        return jsonify({'success': False, 'message': str(e)}), 500
+        return jsonify({'success': False, 'message': f"Internal Server Error: {str(e)}"}), 500
     
 # Endpoint to get user's orders
 @app.route('/myOrder', methods=['GET'])
 def get_user_orders():
     try:
-        user_id = request.args.get('userId', '')
+        user_id = request.args.get('email')
 
+        if user_id == None:
+            return jsonify({"success": False, "message":"Missing Parameter"}),406
+        
         # Create a cursor object to interact with the database
         cursor = myDB.cursor(dictionary=True)
 
@@ -177,23 +160,31 @@ def get_user_orders():
         cursor.execute(query)
         orders = cursor.fetchall()
 
-        # Convert the query result to a list of dictionaries for JSON serialization
-        orders_list = [
-            {
-                'id': order['id'],
-                'totalPrice': float(order['totalPrice']),
-                'orderDate': order['orderDate'].isoformat(),
-                'deliveryDate': order['delevaryDate'].isoformat(),
-                'status': order['status'],
-                'userID': order['userID'],
-            }
-            for order in orders
-        ]
+        for i in range(len(orders)):
+            # Get Order Items
+            query = f"""
+                SELECT 
+                    product.id,
+                    product.productName,
+                    product.productPrice,
+                    product.productImage,
+                    op.quantity as orderedQuantity,
+                    product.calories,
+                    product.discount,
+                    Brand.name as brand,
+                    Brand.nationality 
+                FROM orderProducts as op 
+                    INNER JOIN product on op.product=product.id 
+                    INNER JOIN Brand on product.brand = Brand.id 
+                WHERE `order`={orders[i]['id']};
+            """
+            cursor.execute(query)
+            orders[i]['products'] = cursor.fetchall()
 
         cursor.close()
 
-        return jsonify({'success': True, 'orders': orders_list})
+        return jsonify({'success': True, 'data': orders})
 
     except Exception as e:
         print('Error executing get_user_orders query:', str(e))
-        return jsonify({'success': False, 'error': 'Internal Server Error'}), 500
+        return jsonify({'success': False, 'message': f'Internal Server Error: {str(e)}'}), 500
